Intellij-MicroservicesGuru/mssc-beer-service-2/src/test/java/guru/sfg/msscbrewery/web/controller/BeerControllerTest.java
//package com.vtech.msscbrewery.web.controller;
//
//import static org.mockito.ArgumentMatchers.any;
//import static org.mockito.ArgumentMatchers.anyBoolean;
//import static org.mockito.BDDMockito.given;
//import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
//import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
//import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
//import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
//
//import java.math.BigDecimal;
//import java.util.UUID;
//
//import org.junit.jupiter.api.Test;
//import org.springframework.beans.factory.annotation.Autowired;
//import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
//import org.springframework.boot.test.mock.mockito.MockBean;
//import org.springframework.http.MediaType;
//import org.springframework.test.web.servlet.MockMvc;
//
//import com.fasterxml.jackson.databind.ObjectMapper;
//import com.vtech.msscbrewery.bootstrap.BeerLoader;
//import com.vtech.msscbrewery.services.BeerService;
//import com.vtech.msscbrewery.web.model.BeerDto;
//import com.vtech.msscbrewery.web.model.BeerStyleEnum;
//
//
//
//@WebMvcTest(BeerController.class)
//class BeerControllerTest {
//
//    @Autowired
//    MockMvc mockMvc;
//
//    @Autowired
//    ObjectMapper objectMapper;
//
//    @MockBean
//    BeerService beerService;
//
//    @Test
//    void getBeerById() throws Exception {
//
//        given(beerService.getById(any(), anyBoolean())).willReturn(getValidBeerDto());
//
//        mockMvc.perform(get("/api/v1/beer/" + UUID.randomUUID().toString()).accept(MediaType.APPLICATION_JSON))
//                .andExpect(status().isOk());
//
//    }
//
//    @Test
//    void saveNewBeer() throws Exception {
//
//        BeerDto beerDto = getValidBeerDto();
//        String beerDtoJson = objectMapper.writeValueAsString(beerDto);
//
//        given(beerService.saveNewBeer(any())).willReturn(getValidBeerDto());
//
//        mockMvc.perform(post("/api/v1/beer/")
//                .contentType(MediaType.APPLICATION_JSON)
//                .content(beerDtoJson))
//                .andExpect(status().isCreated());
//    }
//
//    @Test
//    void updateBeerById() throws Exception {
//        given(beerService.updateBeer(any(), any())).willReturn(getValidBeerDto());
//
//        BeerDto beerDto = getValidBeerDto();
//        String beerDtoJson = objectMapper.writeValueAsString(beerDto);
//
//        mockMvc.perform(put("/api/v1/beer/" + UUID.randomUUID().toString())
//                .contentType(MediaType.APPLICATION_JSON)
//                .content(beerDtoJson))
//                .andExpect(status().isNoContent());
//    }
//
//    BeerDto getValidBeerDto(){
//        return BeerDto.builder()
//                .beerName("My Beer")
//                .beerStyle(BeerStyleEnum.ALE)
//                .price(new BigDecimal("2.99"))
//                .upc(BeerLoader.BEER_1_UPC)
//                .build();
//    }
//}